Added unit tests, fixed unit tests
