+++ unittests/scans/anchoreCTL_vulns/many_vulns.json
