+++ unittests/scans/anchoreCTL_vulns/no_vuln.json
@@ -1,0 +1,1 @@
+[]
