+++ unittests/scans/anchoreCTL_vulns/one_vuln.json
@@ -10,6 +10,7 @@
         "packageType": "dpkg",
         "packageVersion": "3.5.8-5+deb9u4",
         "severity": "Medium",
+        "description": "That test description",
         "url": "https://security-tracker.debian.org/tracker/CVE-2011-3389",
         "vuln": "CVE-2011-3389"
     }
