--- unittests/tools/test_anchoreCTL_parser.py
@@ -1,25 +1,0 @@
-from ..dojo_test_case import DojoTestCase
-from dojo.tools.anchorectl.parser import AnchoreCTLParser
-from dojo.models import Test
-
-
-class TestAnchoreCTLParser(DojoTestCase):
-    def test_anchore_engine_parser_has_no_finding(self):
-        testfile = open("unittests/scans/anchoreCTL/no_vuln.json")
-        parser = AnchoreCTLParser()
-        findings = parser.get_findings(testfile, Test())
-        self.assertEqual(0, len(findings))
-
-    def test_anchore_engine_parser_has_one_finding(self):
-        testfile = open("unittests/scans/anchoreCTL/one_vuln.json")
-        parser = AnchoreCTLParser()
-        findings = parser.get_findings(testfile, Test())
-        testfile.close()
-        self.assertEqual(1, len(findings))
-
-    def test_anchore_engine_parser_has_many_findings(self):
-        testfile = open("unittests/scans/anchoreCTL/many_vulns.json")
-        parser = AnchoreCTLParser()
-        findings = parser.get_findings(testfile, Test())
-        testfile.close()
-        self.assertEqual(23, len(findings))
